Simplified templates and fixed acronym
