+++ exercises/practice/acronym/.meta/Sources/Acronym/AcronymExample.swift
@@ -2,7 +2,15 @@ import Foundation
 
 class Acronym {
     static func abbreviate(_ phrase: String) -> String {
-        let words = phrase.components(separatedBy: CharacterSet(charactersIn: " -"))
-        return words.reduce("") { $0 + $1.prefix(1) }
+    let words = phrase.components(separatedBy: CharacterSet(charactersIn: " -_"))
+    var acronym = ""
+    
+    for word in words {
+        if let firstCharacter = word.first {
+            acronym.append(String(firstCharacter).uppercased())
+        }
+    }
+    
+    return acronym
     }
 }
