+++ exercises/practice/acronym/Sources/Acronym/Acronym.swift
@@ -1,16 +1,5 @@
-import Foundation
-
-class Acronym {
-    static func abbreviate(_ phrase: String) -> String {
-    let words = phrase.components(separatedBy: CharacterSet(charactersIn: " -_"))
-    var acronym = ""
-    
-    for word in words {
-        if let firstCharacter = word.first {
-            acronym.append(String(firstCharacter).uppercased())
-        }
-    }
-    
-    return acronym
-    }
+struct Acronym {
+  static func abbreviate(_ inString: String) -> String {
+    // Write your code for the 'Acronym' exercise in this file.
+  }
 }
