+++ generator/Sources/Generator/generator-plugins.swift
@@ -60,7 +60,8 @@ class GeneratorPlugins {
 
     ext.registerFilter("toStringArray") { (value: Any?) in
       if let inputString = value as? [String] {
-        return inputString.joined(separator: "\", \"")
+        guard inputString.isEmpty else { return "[]" }
+        return "\"[\(inputString.joined(separator: "\", \""))\"]"
       }
       return nil
     }
